remove design form for MAB, and combine with metadata form
