+++ src/app/datasources/[datasourceId]/experiments/create/containers/mab/mab-metadata-form.tsx
@@ -14,7 +14,7 @@ import {
   Heading,
 } from '@radix-ui/themes';
 import { PlusIcon, TrashIcon } from '@radix-ui/react-icons';
-import { MABFormData, MABArm } from '@/app/datasources/[datasourceId]/experiments/create/types';
+import { MABFormData, MABArm, PriorType, OutcomeType } from '@/app/datasources/[datasourceId]/experiments/create/types';
 import { NavigationButtons } from '@/components/features/experiments/navigation-buttons';
 import { SectionCard } from '@/components/ui/cards/section-card';
 import { useCreateExperiment } from '@/api/admin';
@@ -29,6 +29,27 @@ interface MABMetadataFormProps {
   onNext: () => void;
   onBack: () => void;
 }
+
+interface OutcomeOption {
+  type: OutcomeType;
+  title: string;
+  description: string;
+}
+
+const OUTCOME_OPTIONS: OutcomeOption[] = [
+  {
+    type: 'binary',
+    title: 'Binary',
+    description:
+      'Yes/No outcomes: conversions, clicks, sign-ups, purchases. Results are expressed as percentages or rates.',
+  },
+  {
+    type: 'real-valued',
+    title: 'Real-valued',
+    description:
+      'Continuous numeric outcomes: revenue per user, time spent, satisfaction scores, any measurable quantity.',
+  },
+];
 
 export function MABMetadataForm({ webhooks, formData, onFormDataChange, onNext, onBack }: MABMetadataFormProps) {
   const { trigger: triggerCreateExperiment, error: createExperimentError } = useCreateExperiment(
@@ -59,6 +80,49 @@ export function MABMetadataForm({ webhooks, formData, onFormDataChange, onNext, 
       ...formData,
       [field]: value,
     });
+  };
+
+  const handleOutcomeChange = (outcomeType: OutcomeType) => {
+    // Auto-map prior type based on outcome type for MAB
+    const priorType: PriorType = outcomeType === 'binary' ? 'beta' : 'normal';
+
+    // Update existing arms to have the correct prior parameters
+    const updatedArms = formData.arms.map((arm) => {
+      const baseArm = {
+        arm_name: arm.arm_name,
+        arm_description: arm.arm_description,
+      };
+
+      if (priorType === 'beta') {
+        // Convert to beta parameters (or use defaults)
+        return {
+          ...baseArm,
+          alpha_prior: arm.alpha_prior || 1,
+          beta_prior: arm.beta_prior || 1,
+          // Remove normal parameters
+          mean_prior: undefined,
+          stddev_prior: undefined,
+        };
+      } else {
+        // Convert to normal parameters (or use defaults)
+        return {
+          ...baseArm,
+          mean_prior: arm.mean_prior ?? 0,
+          stddev_prior: arm.stddev_prior || 1,
+          // Remove beta parameters
+          alpha_prior: undefined,
+          beta_prior: undefined,
+        };
+      }
+    });
+
+    const updatedData = {
+      ...formData,
+      outcomeType,
+      priorType, // Automatically set based on outcome
+      arms: updatedArms, // Update arms with correct parameters
+    };
+    onFormDataChange(updatedData);
   };
 
   const updateArm = (index: number, updatedArm: Partial<MABArm>) => {
@@ -114,8 +178,8 @@ export function MABMetadataForm({ webhooks, formData, onFormDataChange, onNext, 
     const basicValid =
       formData.name.trim() &&
       formData.hypothesis.trim() &&
-      formData.priorType && // Prior type must be selected in design step
-      formData.outcomeType && // Outcome type must be selected in design step
+      formData.priorType &&
+      formData.outcomeType &&
       formData.arms.length >= 2;
 
     const armsValid = formData.arms.every((arm) => arm.arm_name.trim() && isPriorParamValid(arm));
@@ -178,27 +242,49 @@ export function MABMetadataForm({ webhooks, formData, onFormDataChange, onNext, 
         </Flex>
       </SectionCard>
 
+      {/* Outcome Type */}
+      <SectionCard title="Outcome Type">
+        <Text size="2" color="gray" mb="20px">
+          Define the type of outcome measured in this experiment. The prior distribution will be automatically selected
+          based on your choice.
+        </Text>
+        <Text size="2" weight="medium" mb="12px" mt="12px" as="div">
+          Select outcome type
+        </Text>
+
+        <Flex direction="column" gap="3">
+          {OUTCOME_OPTIONS.map((option) => (
+            <OutcomeOptionCard
+              key={option.type}
+              option={option}
+              isSelected={formData.outcomeType === option.type}
+              onSelect={() => handleOutcomeChange(option.type)}
+            />
+          ))}
+        </Flex>
+        {formData.priorType && formData.outcomeType && (
+          <Box p="3">
+            <Flex direction="column" gap="1">
+              <Text size="2" weight="medium" color="blue">
+                Selected Configuration: {formData.priorType === 'beta' ? 'Beta Distribution' : 'Normal Distribution'} ×{' '}
+                {formData.outcomeType === 'binary' ? 'Binary' : 'Real-valued'} Outcome
+              </Text>
+              <Text size="1" color="gray" mt="4px">
+                {formData.priorType === 'beta'
+                  ? 'Using Alpha (prior successes) and Beta (prior failures) parameters'
+                  : 'Using Mean and Standard Deviation parameters'}
+              </Text>
+            </Flex>
+          </Box>
+        )}
+      </SectionCard>
+
       {/* Treatment Arms */}
       <SectionCard title="Arms">
         <Flex direction="column" gap="1">
           <Text size="2" color="gray">
             Configure your experiment arms with prior beliefs.
           </Text>
-          {formData.priorType && formData.outcomeType && (
-            <Box p="3">
-              <Flex direction="column" gap="1">
-                <Text size="2" weight="medium" color="blue">
-                  Selected Configuration: {formData.priorType === 'beta' ? 'Beta Distribution' : 'Normal Distribution'}{' '}
-                  × {formData.outcomeType === 'binary' ? 'Binary' : 'Real-valued'} Outcome
-                </Text>
-                <Text size="1" color="gray" mt="4px">
-                  {formData.priorType === 'beta'
-                    ? 'Using Alpha (prior successes) and Beta (prior failures) parameters'
-                    : 'Using Mean and Standard Deviation parameters'}
-                </Text>
-              </Flex>
-            </Box>
-          )}
           {!formData.priorType ||
             (!formData.outcomeType && (
               <Box p="3">
@@ -263,6 +349,34 @@ export function MABMetadataForm({ webhooks, formData, onFormDataChange, onNext, 
 
       <NavigationButtons onBack={onBack} onNext={handleSaveExperiment} nextLabel="Next" nextDisabled={!isFormValid()} />
     </Flex>
+  );
+}
+
+interface OutcomeOptionCardProps {
+  option: OutcomeOption;
+  isSelected: boolean;
+  onSelect: () => void;
+}
+
+function OutcomeOptionCard({ option, isSelected, onSelect }: OutcomeOptionCardProps) {
+  return (
+    <Card
+      size="2"
+      onClick={onSelect}
+      asChild
+      style={isSelected ? { border: '2px solid var(--accent-9)', backgroundColor: 'var(--accent-2)' } : {}}
+    >
+      <Box as="div">
+        <Flex direction="column" gap="1">
+          <Text size="3" weight="bold">
+            {option.title}
+          </Text>
+          <Text size="2" color="gray">
+            {option.description}
+          </Text>
+        </Flex>
+      </Box>
+    </Card>
   );
 }
